PYTHON/SCRAPPER/projet-scraping-fastapi/python/requirements.txt
# Web Scraping
scrapy==2.11.0
playwright==1.40.0
selenium==4.15.2
beautifulsoup4==4.12.2
lxml==4.9.3

# IA et NLP
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
huggingface-hub==0.19.4

# Email
sendgrid==6.11.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3

# Logging et monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# API server
fastapi==0.115.0
uvicorn==0.30.6
